feat(home): put the live board on the homepage, below the fold
Reverts the lobby-as-homepage flip. `/` is the old homepage again — hero,
Discord CTA, the three mode columns — with live state layered on instead of
replacing it.

Two additions, both chosen to leave the first impression alone:

- The presence line stops saying just "13 players online" and says what those
  people are doing: "13 online · 2 open lobbies · 4 building · 9 solo play".
  A zero is never a line item; an activity that isn't happening is omitted
  rather than shown as 0.
- The board sits under the columns as "Playing right now". Someone who
  scrolls gets open lobbies, pods, and the Karabast games; someone who
  doesn't sees exactly the homepage they've always seen.

The board and everything needed to act on it — join, leave, create, the
Karabast handoff, all three modals — moved into LobbyBoardSection, which /
and /lobby both render. Otherwise two copies of handleCreated (with its
Karabast claim + casual-intent logic) would sit in two files and drift.

The section takes its live data as PROPS rather than calling the hooks
itself. useOpenGamesSocket opens its own socket per call, so a page that also
shows counts in its header — which the homepage now does — would otherwise
hold two connections and fetch the board twice. Verified on the homepage: one
/api/open-games fetch, one /api/pods/public fetch.

usePublicPodsSocket's PublicPod type is exported so the section can name it;
LobbyBoard's local duplicate of that interface is left alone for now.
