+++ src/components/LandingPage.tsx
@@ -1,11 +1,14 @@
 // @ts-nocheck
 'use client'
 
-import { useEffect, useMemo, useRef, useState } from 'react'
+import { useEffect, useMemo, useRef, useState, Fragment } from 'react'
 import { useRouter, useSearchParams } from 'next/navigation'
 import { useAuth } from '../contexts/AuthContext'
 import { usePresence } from '../hooks/usePresence'
 import { usePublicPodsSocket } from '../hooks/usePublicPodsSocket'
+import { useOpenGamesSocket } from '../hooks/useOpenGamesSocket'
+import { useKarabastLobbies } from '../hooks/useKarabastLobbies'
+import { useCompanionCapability } from '../hooks/useCompanionCapability'
 import { formatPoolLabel } from '../utils/poolDisplayName'
 import { getUpcomingSetForPromo } from '../utils/membership'
 import {
@@ -16,6 +19,8 @@ import {
 import { trackEvent, AnalyticsEvents } from '../hooks/useAnalytics'
 import ReleaseNotes from './ReleaseNotes'
 import SiteFooter from './SiteFooter'
+import LobbyBoardSection from './Lobby/LobbyBoardSection'
+import './Lobby/Lobby.css'
 import Button from './Button'
 import SubscribeModal from './SubscribeModal'
 import Countdown from './Countdown'
@@ -120,7 +125,13 @@ function LandingPage() {
       setWasRemoved(true)
     }
   }, [])
-  const { count: playerCount } = usePresence(user?.id)
+  const presence = usePresence(user?.id)
+  const playerCount = presence.count
+  // One instance each, shared with the board below: these hooks each open
+  // their own socket, so calling them twice on one page would double up.
+  const openGames = useOpenGamesSocket()
+  const karabast = useKarabastLobbies()
+  const { casualCapable } = useCompanionCapability()
   const publicPods = usePublicPodsSocket()
   const [activeDraft, setActiveDraft] = useState<ActiveDraft | null>(null)
   const [activeSealedPod, setActiveSealedPod] = useState<ActiveSealedPod | null>(null)
@@ -489,7 +500,21 @@ function LandingPage() {
         {playerCount > 0 && (
           <div className="players-online-landing">
             <span className="online-dot-landing" />
-            <span>{playerCount} player{playerCount !== 1 ? 's' : ''} online</span>
+            {/* A bare total reads as a room full of people refusing to play with
+                you. The breakdown says what they're actually doing, and a zero is
+                never a line item — an omission reads as a gap you can fill. */}
+            <span className="landing-live-line">
+              <strong>{playerCount}</strong> online
+              {[
+                openGames.listings.length > 0 && <><strong>{openGames.listings.length}</strong> open {openGames.listings.length === 1 ? 'lobby' : 'lobbies'}</>,
+                publicPods.length > 0 && <><strong>{publicPods.length}</strong> {publicPods.length === 1 ? 'pod' : 'pods'} forming</>,
+                presence.drafting > 0 && <><strong>{presence.drafting}</strong> drafting</>,
+                presence.building > 0 && <><strong>{presence.building}</strong> building</>,
+                presence.browsing > 0 && <><strong>{presence.browsing}</strong> solo play</>,
+              ].filter(Boolean).map((bit, i) => (
+                <Fragment key={i}> · {bit}</Fragment>
+              ))}
+            </span>
           </div>
         )}
         {activeDraft && (
@@ -614,6 +639,19 @@ function LandingPage() {
           </div>
         </div>
       </div>
+      {/* The live board, below the fold. The hero and the mode columns are
+          untouched — this is what someone finds when they scroll, rather than
+          something competing with the front door. */}
+      <section className="landing-board" aria-label="Open games">
+        <h2 className="landing-board-title">Playing right now</h2>
+        <LobbyBoardSection
+          board={openGames}
+          pods={publicPods}
+          karabast={karabast}
+          companionCapable={casualCapable}
+          returnPath="/"
+        />
+      </section>
       <SiteFooter />
     </div>
   )
